feat (countryList)
organizing home rendering
